FIXED THE FLOODFILL
there was an error with editing the sources during the for loop, this would be a good example 
to show in the science fair. This version has the callstack type stuff where it prints everything out.

issue was with the way i was depreciating the sources, it would remove it from the list and the current
for loop index would get screwed up. I removed the depreciation so it happens outside (before) the for loop in the mapgrid update script.